fix: bug/issue 25 (#94)
* feat: start substitute default settings in ensure_keys

* feat: prompt sub missing settings instead of warn

This is really a final enhancement to the #25 'Handle Missing Keys'
issue. We can now substitute missing settings with default app values,
or we can type a new custom value in, all on initial app startup.

FUTURE CONSIDERATIONS:
If configuration management were to become more important in future,
the following may need to be addressed:

- If a whole section is missing with an inline comment, a different
function parses the diff to the usual. Some of this code could be better
shared with the main function. There's currently minor inconsistency.

- Missing settings are declared as `root['app']['loglevel']`. This could
be more human readable. e.g. `app -> loglevel`

- Expected custom values are not parsed or validated properly. They're
just taken as is.

- If a whole setting section is missing. The default value is an
ordereddict of all the missing nested key/vals within.
This isn't human readable and it's **definitely** not possible to type
out a single custom value to represent that whole structure,
instead each nested key/val within should be iterated one-by-one.

- This would all be necessary if we wanted to add a config command to
the CLI, but surely YAML is easy enough.