ci: terminology change - 'version-constraint key'
Using 'version-constraint key' instead of 'queue-id' to be more descriptive.
Celery terminology is already opaque to end users, i.e. 'routing-key'
vs 'queue'. Queue is both a noun and a verb, and refers to both Celery's
queue name that receives queued tasks as well as the act of queuing them.