refactor(launch_workers.py): logging queue-name reflect term change
Still in a bit of a weird place with terminology here. We have `queue`, `version constraint key` and `routing key`. The choice to refer to the name of the queue as `version constraint key` better reflects application usage and user expectations. "Queue" on its own is both a noun and a verb and we already use the term to describe a module and the process of "queuing" proxies. Some additional terminology changes in code may be necessary at a later date.