fix(ProgressTracker): Polling instead of pub/sub
This really was to get this long-awaited feature out the door.
Issues with messages buffering with Redis pub/sub was preventing
progress from showing accurately. It would slow right down when lots of
tasks were queued. Pub/sub probably needs some multi-threading support
to work properly. Since the buffering is taking place at a TCP level, we
can't just ask Redis to ignore 'guaranteed delivery' (which is not important
for us). This works for now, but I'm no expert on interacting with databases.

Otherwise, the pub/sub pattern worked really nicely for the
following reasons:

- easy channel pattern matching
- channel-based handlers
- messages are only received once, as opposed to
having to ignore the same keys with get/set
- keys aren't stored in Redis once all subscribers have received