perf(ProgressTracker): Store SHA1 for prior data
To avoid reading the same data multiple times, we need to make sure
we can compare new data to stale data. I'm not sure what the best way
to do it is, but we were storing all the progress dicts and comparing
each new dict with the list of old ones.

Surely performance would eventually degrade and storing whole dicts is
not wise. I've replaced storing dicts with storing quick SHA1 sums.
Calculating the sums doesn't seem to affect performance negatively.
Will keep an eye on it. Open to suggestions.